requirements.txt
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
databricks-sdk==0.65.0
docker==7.1.0
fastapi==0.116.2
Flask==3.1.2
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
narwhals==2.5.0
numpy==2.3.3
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.48.0
streamlit==1.49.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
zipp==3.23.0
